+++ tests/data/run/control/08-Test-Run-With-Deletion/03/02/to_process.py
@@ -1,0 +1,5 @@
+def the_answer():
+  """
+  Return the only answer that matters
+  """
+  pass  # YOUR CODE HERE