+++ tests/functional/run/test_run.py
@@ -1,10 +1,14 @@
 
 from wagon_common.helpers.directories import are_directories_identical
+from wagon_common.helpers.file import cp
+from wagon_common.helpers.subprocess import run_command
+
+from wagon_common.helpers.git.create import git_init, git_add, git_commit
 
 from wagon_sync.challengify import Challengify
 from wagon_sync.run_sync import run_sync
 
-import unittest
+import pytest
 
 import os
 import shutil
@@ -12,21 +16,26 @@ import shutil
 from colorama import Fore, Style
 
 
-class TestRun(unittest.TestCase):
+class TestRun():
     """
     test that challenge versions are correctly generated from source codebase
     """
 
+    data_path = os.path.join("tests", "data", "run")
+
+    in_path = os.path.join(data_path, "source")
+    out_path = os.path.join(data_path, "processed")
+    control_path = os.path.join(data_path, "control")
+    seed_path = os.path.join(data_path, "seed")
+
     def test_run(self):
+
+        in_path = os.path.join(self.in_path, "07-Test-Run")
+        out_path = os.path.join(self.out_path, "07-Test-Run")
+        control_path = os.path.join(self.control_path, "07-Test-Run")
 
         # Arrange
         challengify = Challengify()
-
-        data_path = os.path.join("tests", "data", "run")
-
-        in_path = os.path.join(data_path, "source")
-        out_path = os.path.join(data_path, "processed")
-        control_path = os.path.join(data_path, "control")
 
         # Act
         run_sync(
@@ -63,6 +72,78 @@ class TestRun(unittest.TestCase):
         # Cleanup
         shutil.rmtree(out_path, ignore_errors=True)
 
+    @pytest.fixture
+    def deletion_scenario(self):
 
-if __name__ == '__main__':
-    unittest.main()
+        # Arrange
+        challengify = Challengify()
+
+        in_path = os.path.join(self.in_path, "08-Test-Run-With-Deletion")
+        out_path = os.path.join(self.out_path, "08-Test-Run-With-Deletion")
+        control_path = os.path.join(self.control_path, "08-Test-Run-With-Deletion")
+        seed_path = os.path.join(self.seed_path, "08-Test-Run-With-Deletion")
+
+        file_to_delete = os.path.join("03", "02", "to_delete.py")
+
+        # Git init, add, commit in source path
+        git_init(in_path)
+        git_add(in_path)
+        git_commit(in_path, message="Initial commit")
+
+        # Simulate file deletion and git tracking in source path
+        os.remove(os.path.join(in_path, file_to_delete))
+        git_add(in_path)
+
+        # Git init, add, commit in processed path
+        git_init(out_path)
+        git_add(out_path)
+        git_commit(out_path, message="Initial commit")
+
+        yield challengify, file_to_delete, in_path, out_path, control_path
+
+        # Cleanup
+
+        # Put back the "to-be-deleted" file in destination and source
+        cp(os.path.join(seed_path, file_to_delete), os.path.join(in_path, file_to_delete))
+        cp(os.path.join(seed_path, file_to_delete), os.path.join(out_path, file_to_delete))
+
+        # Remove the .git folders in source and destination
+        shutil.rmtree(os.path.join(out_path, ".git"), ignore_errors=True)
+        shutil.rmtree(os.path.join(in_path, ".git"), ignore_errors=True)
+
+    def test_run_with_deletion(self, deletion_scenario):
+
+        challengify, file_to_delete, in_path, out_path, control_path = deletion_scenario
+
+        # Act
+
+        # when no diff is found the whole repo gets synced and the control path content is not valid
+        # git_diff_command = ["git", "diff", "--name-only", "HEAD"]
+        # rc, output, error = run_command(git_diff_command, cwd=in_path)
+        # diff_files = output.decode("utf-8").split()
+        diff_files = [file_to_delete]
+
+        run_sync(
+            challengify=challengify,
+            sources=diff_files,
+            destination=out_path,
+            force=True,
+            dry_run=False,
+            verbose=True,
+            test=False,
+            ignore_tld=True,
+            iterate_yaml_path=".")
+
+        # Assert
+        rc, output, error = are_directories_identical(out_path, control_path, ignored_files=[".git"])
+
+        if rc != 0:
+            print(Fore.RED
+                  + "\nDirectory content does not match 🤕"
+                  + Style.RESET_ALL
+                  + f"\n- rc: {rc}"
+                  + f"\n- output: {output}"
+                  + f"\n- error: {error}")
+            print(output.decode("utf-8"))
+
+        assert rc == 0