+++ wagon_sync/params/home_defaults.py
@@ -1,4 +1,6 @@
 
+from wagon_common.helpers.git.repo import get_git_top_level_directory
+
 import os
 import yaml
 
@@ -102,7 +104,14 @@ def load_home_defaults(verbose):
 
     run_defaults = defaults_conf.get(CMD_RUN, {})
 
-    run_destination_default = run_defaults.get(CMD_RUN_DESTINATION, CMD_RUN_DESTINATION_DEFAULT)
+    # process solutions repo default destination
+    cmd_run_destination_default = CMD_RUN_DESTINATION_DEFAULT
+    if CMD_RUN_DESTINATION not in run_defaults:
+        tld = get_git_top_level_directory()
+        if tld[-len("-solutions"):] == "-solutions":
+            cmd_run_destination_default = tld[:-len("-solutions")] + "-challenges"
+
+    run_destination_default = run_defaults.get(CMD_RUN_DESTINATION, cmd_run_destination_default)
     run_ignore_cwd_default = run_defaults.get(CMD_RUN_IGNORE_CWD, CMD_RUN_IGNORE_CWD_DEFAULT)
     run_force_default = run_defaults.get(CMD_RUN_FORCE, CMD_RUN_FORCE_DEFAULT)
     run_dry_run_default = run_defaults.get(CMD_RUN_DRY_RUN, CMD_RUN_DRY_RUN_DEFAULT)