+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+
+# 0.3.9 2022-11-25
+
+### Added
+
+- Adds support for file deletion: remove from the destination the input sources that shoud not exist
+- Adds repository related default destination parameter: a `*-solutions` respository will have a `*-challenges` default destination parameter
+
 # 0.3.8 2022-07-01
 
 ### Added